@@ -6,14 +6,7 @@ import unittest
 
 import opentimelineio as otio
 
-
-# handle python2 vs python3 difference
-try:
-    from tempfile import TemporaryDirectory  # noqa: F401
-    import tempfile
-except ImportError:
-    # XXX: python2.7 only
-    from backports import tempfile
+import tempfile
 
 # Reference data
 SAMPLE_DATA_DIR = os.path.join(os.path.dirname(__file__), "sample_data")
